+++ Code/Backend/Backend/Repositories/ICategoryRepository.cs
@@ -4,9 +4,9 @@ namespace Backend.Repositories;
 
 public interface ICategoryRepository
 {
-    Task<Category?> GetByIdAsync(long categoryId);
+    Task<Category?> GetByIdAsync(int categoryId);
     Task<List<Category>> GetAllAsync();
-    Task<List<Category>> GetChildrenAsync(long parentId);
+    Task<List<Category>> GetChildrenAsync(int parentId);
     Task<List<Category>> GetRootCategoriesAsync();
     Task AddAsync(Category category);
     void Update(Category category);
